requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-storage==2.10.0
httpx==0.25.2
apscheduler==3.10.4
requests==2.31.0 